Refactor server/worker out of client module (#846)
* Refactor server/worker out of client

* Fix imports
