+++ turbinia/client_test.py
@@ -20,23 +20,14 @@ from datetime import datetime
 from datetime import timedelta
 import json
 import unittest
-import os
-import shutil
-import tempfile
 import textwrap
 
 import mock
 
 from turbinia import config
 from turbinia import client as TurbiniaClientProvider
-from turbinia.client import TurbiniaServer
-from turbinia.client import TurbiniaStats
-from turbinia.client import TurbiniaPsqWorker
-from turbinia.client import check_system_dependencies
-from turbinia.client import check_docker_dependencies
-from turbinia.jobs import manager
-from turbinia.jobs import manager_test
 from turbinia import TurbiniaException
+from turbinia.client import TurbiniaStats
 
 DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%S.%fZ'
 
@@ -589,170 +580,3 @@ class TestTurbiniaStats(unittest.TestCase):
     stats.calculate_stats()
     report = stats.format_stats_csv()
     self.assertEqual(report, test_output)
-
-
-class TestTurbiniaServer(unittest.TestCase):
-  """Test Turbinia Server class."""
-
-  @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
-  def testTurbiniaServerInit(self, _, __):
-    """Basic test for Turbinia Server init."""
-    server = TurbiniaServer()
-    self.assertTrue(hasattr(server, 'task_manager'))
-
-
-class TestTurbiniaPsqWorker(unittest.TestCase):
-  """Test Turbinia PSQ Worker class."""
-
-  def setUp(self):
-    self.tmp_dir = tempfile.mkdtemp(prefix='turbinia-test')
-    config.LoadConfig()
-    config.OUTPUT_DIR = self.tmp_dir
-    config.MOUNT_DIR_PREFIX = self.tmp_dir
-    config.ParseDependencies = mock.MagicMock(return_value={})
-    self.saved_jobs = manager.JobsManager._job_classes
-
-  def tearDown(self):
-    manager.JobsManager._job_classes = self.saved_jobs
-    if 'turbinia-test' in self.tmp_dir:
-      shutil.rmtree(self.tmp_dir)
-
-  @mock.patch('turbinia.client.pubsub')
-  @mock.patch('turbinia.client.datastore.Client')
-  @mock.patch('turbinia.client.psq.Worker')
-  @mock.patch('turbinia.lib.docker_manager.DockerManager')
-  def testTurbiniaPsqWorkerInit(self, _, __, ___, ____):
-    """Basic test for PSQ worker."""
-    worker = TurbiniaPsqWorker([], [])
-    self.assertTrue(hasattr(worker, 'worker'))
-
-  @mock.patch('turbinia.client.pubsub')
-  @mock.patch('turbinia.client.datastore.Client')
-  @mock.patch('turbinia.client.psq.Worker')
-  @mock.patch('turbinia.lib.docker_manager.DockerManager')
-  def testTurbiniaClientNoDir(self, _, __, ___, ____):
-    """Test that OUTPUT_DIR path is created."""
-    config.OUTPUT_DIR = os.path.join(self.tmp_dir, 'no_such_dir')
-    TurbiniaPsqWorker([], [])
-    self.assertTrue(os.path.exists(config.OUTPUT_DIR))
-
-  @mock.patch('turbinia.client.pubsub')
-  @mock.patch('turbinia.client.datastore.Client')
-  @mock.patch('turbinia.client.psq.Worker')
-  @mock.patch('turbinia.lib.docker_manager.DockerManager')
-  def testTurbiniaClientIsNonDir(self, _, __, ___, ____):
-    """Test that OUTPUT_DIR does not point to an existing non-directory."""
-    config.OUTPUT_DIR = os.path.join(self.tmp_dir, 'empty_file')
-    open(config.OUTPUT_DIR, 'a').close()
-    self.assertRaises(TurbiniaException, TurbiniaPsqWorker)
-
-  @mock.patch('turbinia.client.config')
-  @mock.patch('turbinia.client.check_directory')
-  @mock.patch('turbinia.client.pubsub')
-  @mock.patch('turbinia.client.datastore.Client')
-  @mock.patch('turbinia.client.psq.Worker')
-  @mock.patch('turbinia.lib.docker_manager.DockerManager')
-  def testTurbiniaClientJobsLists(self, _, __, ___, ____, _____, mock_config):
-    """Test that client job allowlist and denylists are setup correctly."""
-    mock_config.PSQ_TOPIC = 'foo'
-    manager.JobsManager._job_classes = {}
-    manager.JobsManager.RegisterJob(manager_test.TestJob1)
-    manager.JobsManager.RegisterJob(manager_test.TestJob2)
-    manager.JobsManager.RegisterJob(manager_test.TestJob3)
-
-    # Check denylist
-    TurbiniaPsqWorker(['testjob1'], [])
-    self.assertListEqual(
-        sorted(list(manager.JobsManager.GetJobNames())),
-        ['testjob2', 'testjob3'])
-    manager.JobsManager.RegisterJob(manager_test.TestJob1)
-
-    # Check denylist with DISABLED_JOBS config
-    mock_config.DISABLED_JOBS = ['testjob1']
-    TurbiniaPsqWorker(['testjob2'], [])
-    self.assertListEqual(list(manager.JobsManager.GetJobNames()), ['testjob3'])
-    manager.JobsManager.RegisterJob(manager_test.TestJob1)
-    manager.JobsManager.RegisterJob(manager_test.TestJob2)
-    mock_config.DISABLED_JOBS = ['']
-
-    # Check allowlist
-    TurbiniaPsqWorker([], ['testjob1'])
-    self.assertListEqual(list(manager.JobsManager.GetJobNames()), ['testjob1'])
-    manager.JobsManager.RegisterJob(manager_test.TestJob2)
-    manager.JobsManager.RegisterJob(manager_test.TestJob3)
-
-    # Check allowlist of item in DISABLED_JOBS config
-    mock_config.DISABLED_JOBS = ['testjob1', 'testjob2']
-    TurbiniaPsqWorker([], ['testjob1'])
-    self.assertListEqual(list(manager.JobsManager.GetJobNames()), ['testjob1'])
-    manager.JobsManager.RegisterJob(manager_test.TestJob2)
-    manager.JobsManager.RegisterJob(manager_test.TestJob3)
-
-  @mock.patch('turbinia.client.subprocess.Popen')
-  @mock.patch('logging.Logger.warning')
-  def testSystemDependencyCheck(self, mock_logger, popen_mock):
-    """Test system dependency check."""
-    dependencies = {
-        'plasojob': {
-            'programs': ['non_exist'],
-            'docker_image': None
-        }
-    }
-    # Dependency not found.
-    proc_mock = mock.MagicMock()
-    proc_mock.communicate.return_value = (b'no', b'thing')
-    proc_mock.returncode = 1
-    popen_mock.return_value = proc_mock
-    self.assertRaises(
-        TurbiniaException, check_system_dependencies, dependencies)
-
-    # Normal run.
-    proc_mock.returncode = 0
-    check_system_dependencies(dependencies)
-
-    # Job not found.
-    dependencies['non_exist'] = dependencies.pop('plasojob')
-    check_system_dependencies(dependencies)
-    mock_logger.assert_called_with(
-        'The job non_exist was not found or has been disabled. '
-        'Skipping dependency check...')
-
-  @mock.patch('turbinia.lib.docker_manager.DockerManager')
-  @mock.patch('turbinia.lib.docker_manager.ContainerManager')
-  @mock.patch('logging.Logger.warning')
-  def testDockerDependencyCheck(
-      self, mock_logger, mock_contmgr, mock_dockermgr):
-    """Test Docker dependency check."""
-    dependencies = {
-        'plasojob': {
-            'programs': ['non_exist'],
-            'docker_image': 'test_img'
-        }
-    }
-
-    # Set up mock objects
-    mock_dm = mock_dockermgr.return_value
-    mock_dm.list_images.return_value = ['test_img']
-    mock_cm = mock_contmgr.return_value
-
-    # Dependency not found.
-    mock_cm.execute_container.return_value = ['non_exist', None, 1]
-    self.assertRaises(
-        TurbiniaException, check_docker_dependencies, dependencies)
-
-    # Normal run
-    mock_cm.execute_container.return_value = ['exists', None, 0]
-    check_docker_dependencies(dependencies)
-
-    # Docker image not found
-    mock_dm.list_images.return_value = ['non_exist']
-    self.assertRaises(
-        TurbiniaException, check_docker_dependencies, dependencies)
-
-    # Job not found.
-    dependencies['non_exist'] = dependencies.pop('plasojob')
-    check_docker_dependencies(dependencies)
-    mock_logger.assert_called_with(
-        'The job non_exist was not found or has been disabled. '
-        'Skipping dependency check...')
